Updated README.md with Additional Documentation for Upgrading for Assetylene 1.x to 2.x
In order to make it easier for Movable Type users to upgrade to Movable Type 6.5, After6 Services updated the README.md to include detailed instructions about updating the Asset Insertion template from Assetylene 1.x to one that is compatible with Assetylene 2.x.

We also standardized the structure of the README to correspond with the format that we have used with other plugins.

After6 Services assumed maintenance of this plugin in mid-2019.