Updated Link to Installation Instructions
Replaced the link to http://tinyurl.com/easy-plugin-install with http://after6services.com/2020/01/the-ultimate-guide-to-installing-movable-type-plugins/. This is because the page referred to by http://tinyurl.com/easy-plugin-install on majordojo.com is no longer accessible; Majordojo.com is currently down.

After6 recreated the "The Ultimate Guide to Installing Movable Type Plugins" page on its website, with permission of its original author, Byrne Reese.